requirements.txt
fastapi==0.119.1
uvicorn[standard]==0.38.0
gunicorn==23.0.0
sqlalchemy==2.0.44
psycopg2-binary==2.9.11
python-jose[cryptography]==3.5.0
bcrypt==4.2.1
pydantic[email]==2.12.3
httpx==0.28.1
python-multipart==0.0.20
celery==5.4.0
redis==5.2.1
cryptography
markdown2
python-docx
reportlab
Pillow
replicate
requests
replit-object-storage
google-cloud-storage
tomli
anthropic
openai
faker
langchain-experimental
presidio-analyzer
presidio-anonymizer
spacy
